style:: 게시글 목록 조회/반환 dto 이름 변경
